# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FinTalent Testing API
  version: 1.0.0
  description: API создания, публикации, прохождения и модерации профессиональных тестов.
servers:
  - url: http://localhost:8080
...
paths:
  /api/tests:
    get:
      tags: [Tests]
      summary: Список доступных или авторских тестов
      parameters:
        - {in: query, name: scope, schema: {type: string, enum: [available, mine]}}
        - {in: query, name: status, schema: {type: string}}
        - {in: query, name: category, schema: {type: string}}
        - {in: query, name: q, schema: {type: string}}
      responses: {'200': {description: Список тестов}}
    post:
      tags: [Tests]
      summary: Создать тест-черновик
      requestBody: {$ref: '#/components/requestBodies/Test'}
      responses: {'201': {description: Тест создан}}
...
components:
...
  requestBodies:
    Test:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/TestInput'}}}
  schemas:
    TestInput:
      type: object
      required: [title]
      properties:
        title: {type: string}
        description: {type: string}
        category: {type: string}
        difficulty: {type: string, enum: [easy, medium, hard]}
        visibility: {type: string, enum: [private, public, marketplace]}
        is_free: {type: boolean}
        price: {type: number}
        passing_percent: {type: number, minimum: 0, maximum: 100}
        time_limit_seconds: {type: integer, nullable: true}
...